ci: unblock fork PRs and run the core suite before the permission gate
