+++ .github/workflows/CONTEXT.md
@@ -2,7 +2,7 @@
 
 CI and release machinery. Read the YAML for mechanics; this covers what's *not* visible there — the structure, the security model, and the publishing rules that bite.
 
-Dependabot (config, the alerts-vs-PRs distinction, and the deliberate-pin ledger) is documented one level up in [`../CONTEXT.md`](../CONTEXT.md) and [`DEPENDENCY-MANAGEMENT.md`](../../DEPENDENCY-MANAGEMENT.md).
+Dependabot (config, the alerts-vs-PRs distinction, and the deliberate-pin ledger) is documented one level up in [`../CONTEXT.md`](../CONTEXT.md) and [`docs/dependency-management/CONTEXT.md`](../../docs/dependency-management/CONTEXT.md).
 
 ## CI
 
@@ -10,12 +10,78 @@ Dependabot (config, the alerts-vs-PRs distinction, and the deliberate-pin ledger
 
 `scripts/ci-test-sanity-check.sh` (run by `main.yaml`) fails if any `*.spec.ts(x)` isn't wired into a `jest.config.ts` project — so no test can be silently absent from CI.
 
-### Safely accepting external PRs — do not break this
+### The design: why external-PR CI is shaped this way — do not break this
 
-`run-tests.yaml` triggers on **`pull_request_target`**, so it runs with the repo's secrets *and* checks out the PR's head code — arbitrary code from any author running on a runner that holds production credentials. This is a known privilege-escalation footgun; the **`check-permission`** job (`malloydata/check-ci-permissions`) contains it, failing the run unless **`github.triggering_actor`** has write access (design: PR #2087).
+The whole shape follows from one platform constraint. Malloy's dialect tests need live
+warehouse credentials (BigQuery, Snowflake, Databricks, and the BigQuery-backed
+Trino/Presto containers). **There is exactly one trigger under which a pull request from a
+fork can reach a repository secret, and it is `pull_request_target`.**
 
-- **Every secret-bearing job MUST `needs: check-permission`** (today: `main`, `db-trino`, `db-presto`, `db-bigquery`, `db-snowflake`, `db-databricks`). Secret-free jobs deliberately don't. Adding a secret to an ungated job leaks it to external-PR code.
-- The gate trusts the *triggering actor*, not the PR author — so **re-running a fork PR's CI authorizes its code to run with full secrets.** Review the diff (especially workflow/build-script changes) before re-running.
+| trigger | PR from | workflow file taken from | secrets | `GITHUB_TOKEN` | GitHub's fork-approval gate |
+|---|---|---|---|---|---|
+| `pull_request` | branch in this repo | PR head | yes | per `permissions:` | no |
+| `pull_request` | **fork** | PR head | **none — `${{ secrets.X }}` is empty** | **read-only, hard cap** | **yes** |
+| `pull_request_target` | branch in this repo | base | yes | per `permissions:` | no |
+| `pull_request_target` | **fork** | **base** | **yes** | per `permissions:`, base-repo scoped | **no** |
+
+That bottom row is the entire security problem and the entire reason the rest exists. It is
+not a preference: on `pull_request` from a fork, secrets are unavailable as a platform
+guarantee, with no setting that turns them on. So a maintainer clicking GitHub's built-in
+"Approve and run" still produces a run whose database tests fail for want of credentials —
+which is what happened, and was reverted, in PR #2078.
+
+`pull_request_target` buys the secrets and costs two things. The workflow comes from the
+**base** branch (good: a fork cannot rewrite the CI definition), but it runs **fork code**
+on a runner holding production credentials, and it is **exempt from GitHub's fork-approval
+gate**. That exemption is why **`check-permission`** exists (`malloydata/check-ci-permissions`,
+design: PR #2087) — it is a hand-built replacement for a control the platform declines to
+apply to this trigger. Note the repo's fork-approval policy is set to
+`all_external_contributors`, the strictest available; it is simply inert here.
+
+**How the gate actually behaves.** `check-permission` does not pause for approval — it
+*fails* unless `github.triggering_actor` has write access, and failing skips every job that
+`needs:` it. The maintainer action is pressing **"Re-run failed jobs"** in the Actions UI,
+which starts a run whose triggering actor is the maintainer. Consequences worth knowing:
+
+- The vouch is "who pressed the button", not "what was reviewed". **Review the diff —
+  especially workflow and build-script changes — before re-running.**
+- A re-run replays the **original event payload**, so the head SHA is fixed; a re-run cannot
+  pick up newer commits. The vouch is pinned to a commit.
+- A new push creates a **new, unvouched run** rather than un-vouching the old one. The
+  hazard is clicking the newest run in the list, which may be a commit you did not read.
+  (A Dependabot rebase does this silently — see the alerts section in [`../CONTEXT.md`](../CONTEXT.md).)
+- If `pull_and_build`'s artifact is still alive (`retention-days: 1`), "Re-run failed jobs"
+  reuses it rather than rebuilding.
+
+**The invariant to preserve: a job may run before the gate if and only if no secrets are
+passed to it.** Stated the old way — every secret-bearing job MUST `needs: check-permission`
+(today: `db-trino`, `db-presto`, `db-bigquery`, `db-snowflake`, `db-databricks`) — but the
+inverted form is the one that generalizes, and it is checkable by reading one file. What an
+external PR can do before the gate is then describable without auditing a single test
+script: spend runner compute and network egress. Nothing else.
+
+**Every job holds two credentials, and you only chose one.** The secrets you pass, and the
+`GITHUB_TOKEN` GitHub injects whether you asked or not. Under `pull_request_target` that
+token is scoped to **this** repo, so fork code could use it to push to `main` if it had any
+power. `permissions: {}` on `run-tests.yaml` and on every reusable workflow sets it to
+none — a called workflow can lower its caller's ceiling but never raise it. **This
+discipline is load-bearing, not decorative: the repo's `default_workflow_permissions` is
+`write`, so a workflow that simply omits `permissions:` gets a write token.**
+
+**`main` is deliberately pre-gate.** It carries no secrets, so external contributors get
+lint and the ~3000 dialect-agnostic tests immediately, without waiting for a maintainer —
+that is the point, and it is most of the value external PR CI provides. It stays that way
+only while `ci-core` is pinned: the script sets `MALLOY_DATABASES=duckdb,postgres` so specs
+cannot fall back to the default database lists hardcoded in their own source. Unpinned,
+`ci-core` is the **only** `ci-*` script that lets such a default select a credentialed
+warehouse — which is exactly how `main` came to require `BIGQUERY_KEY` for a single spec
+file. Both halves matter and catch different mistakes: the pin neutralizes the
+env-driven route, and the absent credential catches a hardcoded one (a literal
+`bigquery.table(...)` in a core spec), which fails loudly rather than quietly spending a
+key, because the harness treats an unreachable connection as an error and not a skip. That
+tripwire assumes the runner carries no ambient cloud credential — true for GitHub-hosted
+runners, and something that would silently void on self-hosted runners with an attached
+service account.
 
 ### The security stance on pull_and_build, stated honestly
